requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
cffi==1.10.0
cryptography==1.8.1
enum34==1.1.6
Fabric==1.13.2
idna==2.5
ipaddress==1.0.18
packaging==16.8
paramiko==2.1.2
pkg-resources==0.0.0
pyaml==16.12.2
pyasn1==0.2.3
pycparser==2.17
PyYAML==3.12
six==1.10.0
